requirements.txt
# Core dependencies
streamlit==1.41.0
PyPDF2==3.0.1
ebooklib==0.20
beautifulsoup4==4.12.3
deep-translator==1.11.4
gTTS==2.5.4
pydub==0.25.1
python-magic==0.4.27
nltk==3.9.1
pygame>=2.5.0

# Performance and GPU dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0

# Basic TTS dependencies (Python 3.13 compatible)
soundfile>=0.12.0
librosa>=0.10.0

# Async and parallel processing
aiohttp>=3.8.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=10.0.0

# Optional enhanced TTS (install separately if needed)
# Note: The following may have Python 3.13 compatibility issues:
# pip install TTS --no-deps --force-reinstall
# pip install coqui-tts --no-deps --force-reinstall

# Optional: For GPU support, install CUDA version separately:
# For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
